breaking live chat into components to be reused in other contexts
